+++ ms-reservas/src/main/java/com/eventos/ms_reservas/exception/NoDisponibleNotFoundException.java
@@ -1,14 +1,14 @@
 package com.eventos.ms_reservas.exception;
 
 public class NoDisponibleNotFoundException extends RuntimeException {
-    private final Long id;
+    private final Integer id;
 
-    public NoDisponibleNotFoundException(Long id, String message) {
+    public NoDisponibleNotFoundException(Integer id, String message) {
         super(message);
         this.id = id;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 }
